fix(cuda): sanitize sharedMemPerBlockOptin for Blackwell SM120
NVIDIA driver bug on early CUDA 13.x returns bogus 0x100000001 for
sharedMemPerBlockOptin on SM120 GPUs. When this size_t value is passed
to cudaFuncSetAttribute (which takes int), it silently truncates to 1,
making any kernel launch with >1 byte dynamic shared memory fail with
cudaErrorInvalidValue.

Two-layer fix:
- ggml-cuda.cu: sanity-check sharedMemPerBlockOptin, fall back to
  sharedMemPerBlock if >256 KB (no real GPU exceeds ~228 KB).
- common.cuh: explicit (int) cast in CUDA_SET_SHARED_MEMORY_LIMIT to
  prevent silent truncation at the call site.

Verified on RTX 5060 Ti (SM120) where mm_ids_helper<4> previously
crashed with cudaErrorInvalidValue for n_tokens>=16, and on RTX 3080
(SM86) with no regression.

+++ ggml/src/ggml-cuda/common.cuh
@@ -233,7 +233,7 @@ static const char * cu_get_error_str(CUresult err) {
             static bool shared_memory_limit_raised[GGML_CUDA_MAX_DEVICES] = { false };                         \
             const int   id                                                = ggml_cuda_get_device();            \
             if (!shared_memory_limit_raised[id]) {                                                             \
-                CUDA_CHECK(cudaFuncSetAttribute(kernel, cudaFuncAttributeMaxDynamicSharedMemorySize, nbytes)); \
+                CUDA_CHECK(cudaFuncSetAttribute(kernel, cudaFuncAttributeMaxDynamicSharedMemorySize, (int) (nbytes))); \
                 shared_memory_limit_raised[id] = true;                                                         \
             }                                                                                                  \
         } while (0)
